Move factory methods to Decorator
